# refactor(Schema) only add the RescueMiddleware if it's being used