Save the rainforest, etc. Middleware is pretty slow.

```ruby 
# ~/code/graphql $ cat middleware_benchmark.rb
require "benchmark/ips"
require "graphql"
require "support/dairy_app"

query = GraphQL::Introspection::INTROSPECTION_QUERY

Benchmark.ips do |x|
  x.report("without empty middleware") { DummySchema.execute(query) }
  x.report("with empty middleware") { DummyWithRescueSchema.execute(query) }
  x.compare!
end
```

```
~/code/graphql $ ruby -Ilib -Ispec middleware_benchmark.rb
Warming up --------------------------------------
without empty middleware
                         3.000  i/100ms
with empty middleware
                         2.000  i/100ms
Calculating -------------------------------------
without empty middleware
                         37.813  (±10.6%) i/s -    186.000  in   5.014397s
with empty middleware
                         21.016  (±23.8%) i/s -    100.000  in   5.036850s

Comparison:
without empty middleware:       37.8 i/s
with empty middleware:       21.0 i/s - 1.80x  slower
```